+++ CHANGELOG.md
@@ -17,7 +17,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - A major rewrite. TrueType tables are no longer public.
 - Use `GlyphId` instead of `u16`.
 
-### Fixed
+### Removed
 - `GDEF` table parsing.
+
+### Fixed
+
+- Panic during a glyph outlining when contour has only one point.
 
 [Unreleased]: https://github.com/RazrFalcon/ttf-parser/compare/v0.1.0...HEAD